MCI.Standard を作成
