+++ src/MCI.Core/MCI.Core.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard1.1</TargetFramework>
     <RootNamespace>Miharu</RootNamespace>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <AssemblyVersion>0.10.0.0</AssemblyVersion>
@@ -10,5 +10,9 @@
     <Company>Miharu Communications Inc.</Company>
     <Authors>Miharu Communications Inc.</Authors>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Net\Sockets\" />
+  </ItemGroup>
 
 </Project>
