--- src/MCI.Core/Net/Sockets/IAsyncSocket.cs
@@ -1,25 +1,0 @@
-namespace Miharu.Net.Sockets
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Net;
-    using System.Net.Sockets;
-    using System.Text;
-    using System.Threading.Tasks;
-
-    public interface IAsyncSocket : IDisposable
-    {
-        ProtocolType Protocol { get; }
-        IPEndPoint RemoteEndPoint { get; }
-
-        int SendTimeout { get; }
-        int ReceiveTimeout { get; }
-        int BufferSize { get; }
-
-        Task<Either<IFailedReason, Unit>> ConnectAsync();
-        Task<Either<IFailedReason, Unit>> DisconnectAsync();
-
-        Task<Either<IFailedReason, Unit>> SendAsync(byte[] packet);
-        Task<Either<IFailedReason, byte[]>> ReceiveAsync();
-    }
-}
