+++ src/MCI.Standard.Tests/Net/IPingService.cs
