+++ src/MCI.Standard.Tests/Net/PingService.cs
