+++ src/MCI.Standard.Tests/Net/Sockets/ReceivedEventArgs.cs
